+++ lib/plugins/aws/logs.js
@@ -4,8 +4,8 @@ const BbPromise = require('bluebird');
 const _ = require('lodash');
 const dayjs = require('dayjs');
 const utc = require('dayjs/plugin/utc');
-const validate = require('../lib/validate');
-const formatLambdaLogEvent = require('../utils/formatLambdaLogEvent');
+const validate = require('./lib/validate');
+const formatLambdaLogEvent = require('./utils/formatLambdaLogEvent');
 
 dayjs.extend(utc);
