+++ lib/plugins/index.js
@@ -27,7 +27,7 @@ module.exports = [
   require('./aws/deploy/index.js'),
   require('./aws/invoke.js'),
   require('./aws/info/index.js'),
-  require('./aws/logs/index.js'),
+  require('./aws/logs.js'),
   require('./aws/metrics/awsMetrics.js'),
   require('./aws/remove/index.js'),
   require('./aws/rollback/index.js'),
