+++ test/unit/lib/plugins/aws/logs.test.js
@@ -3,9 +3,9 @@
 const bluebird = require('bluebird');
 const expect = require('chai').expect;
 const sinon = require('sinon');
-const AwsProvider = require('../../../../../../lib/plugins/aws/provider/awsProvider');
-const AwsLogs = require('../../../../../../lib/plugins/aws/logs/index');
-const Serverless = require('../../../../../../lib/Serverless');
+const AwsProvider = require('../../../../../lib/plugins/aws/provider/awsProvider');
+const AwsLogs = require('../../../../../lib/plugins/aws/logs');
+const Serverless = require('../../../../../lib/Serverless');
 
 describe('AwsLogs', () => {
   let serverless;
